# feat: Complete OpenAI Responses API integration with 62.7% performance improvement