--- a/backend/core/config.py
+++ b/backend/core/config.py
@@ -31,10 +31,9 @@ class Settings(BaseSettings):
     event_batch_size: int = 100
     event_block_ms: int = 1000  # How long to block waiting for events
     
-    # LocoNav
-    loconav_api_key: Optional[str] = None
-    loconav_api_secret: Optional[str] = None
-    loconav_base_url: str = "https://api.loconav.com"
+    # LocoNav - CORRECTED IMPLEMENTATION
+    loconav_user_token: Optional[str] = None
+    loconav_base_url: str = "https://api.a.loconav.com"
     loconav_timeout: int = 30
     
     # Google Sheets