--- a/backend/services/__init__.py
+++ b/backend/services/__init__.py
@@ -3,11 +3,20 @@ Service layer for Watch Tower business logic
 """
 
 from .loconav import LocoNavService
+from .loconav_service import LocoNavAPIService, loconav_api_service
 from .google_sheets import GoogleSheetsService
 from .analytics import AnalyticsService
+from .slack_service import SlackService, slack_service
+from .ai_service import AIService, ai_service
 
 __all__ = [
     "LocoNavService",
+    "LocoNavAPIService", 
+    "loconav_api_service",
     "GoogleSheetsService", 
-    "AnalyticsService"
+    "AnalyticsService",
+    "SlackService",
+    "slack_service",
+    "AIService",
+    "ai_service"
 ]