--- a/backend/tasks/__init__.py
+++ b/backend/tasks/__init__.py
@@ -1,0 +1,16 @@
+"""
+Background tasks for Watch Tower
+"""
+
+from .sync_tasks import sync_google_sheets_data, cleanup_old_positions
+from .analytics_tasks import generate_daily_analytics, check_vehicle_connectivity, generate_weekly_report
+from .notification_tasks import send_daily_summary
+
+__all__ = [
+    "sync_google_sheets_data",
+    "cleanup_old_positions", 
+    "generate_daily_analytics",
+    "check_vehicle_connectivity",
+    "generate_weekly_report",
+    "send_daily_summary"
+]