--- a/backend/tasks/analytics_tasks.py
+++ b/backend/tasks/analytics_tasks.py
@@ -1,0 +1,163 @@
+"""
+Analytics and reporting tasks
+"""
+
+import logging
+from datetime import datetime, timedelta
+from celery import current_app as celery_app
+from services import AnalyticsService, slack_service
+from core.config import settings
+
+logger = logging.getLogger(__name__)
+
+
+@celery_app.task(bind=True, max_retries=2)
+def generate_daily_analytics(self):
+    """Generate daily fleet analytics"""
+    try:
+        logger.info("Starting daily analytics generation")
+        
+        yesterday = datetime.utcnow().date() - timedelta(days=1)
+        
+        analytics_service = AnalyticsService()
+        
+        # In real implementation, this would:
+        # 1. Calculate daily metrics (trips, distance, utilization)
+        # 2. Identify patterns and anomalies
+        # 3. Store results in database
+        # 4. Update Google Sheets dashboard
+        
+        metrics = {
+            "date": yesterday.isoformat(),
+            "total_trips": 0,  # Placeholder
+            "total_distance_km": 0.0,
+            "active_trucks": 0,
+            "completed_trips": 0,
+            "average_trip_time": 0.0,
+            "fuel_efficiency": 0.0
+        }
+        
+        logger.info(f"Daily analytics generated for {yesterday}")
+        return {"status": "success", "metrics": metrics}
+        
+    except Exception as e:
+        logger.error(f"Daily analytics generation failed: {e}")
+        raise self.retry(countdown=300)
+
+
+@celery_app.task(bind=True)
+def check_vehicle_connectivity(self):
+    """Check which vehicles are reporting location data"""
+    try:
+        logger.info("Checking vehicle connectivity")
+        
+        # In real implementation:
+        # 1. Check last position update for each truck
+        # 2. Identify vehicles that haven't reported in > 30 minutes
+        # 3. Send alerts for disconnected vehicles
+        # 4. Update vehicle status
+        
+        disconnected_vehicles = []  # Placeholder
+        
+        if disconnected_vehicles:
+            # Send alert about disconnected vehicles
+            logger.warning(f"Found {len(disconnected_vehicles)} disconnected vehicles")
+            
+            # Send Slack notification if configured
+            if settings.slack_bot_token:
+                # Would call slack_service.send_alert()
+                pass
+        
+        logger.info("Vehicle connectivity check completed")
+        return {
+            "status": "success", 
+            "disconnected_count": len(disconnected_vehicles),
+            "disconnected_vehicles": disconnected_vehicles
+        }
+        
+    except Exception as e:
+        logger.error(f"Vehicle connectivity check failed: {e}")
+        raise
+
+
+@celery_app.task(bind=True, max_retries=2)
+def generate_weekly_report(self):
+    """Generate weekly fleet performance report"""
+    try:
+        logger.info("Starting weekly report generation")
+        
+        # Calculate week ending yesterday
+        end_date = datetime.utcnow().date() - timedelta(days=1)
+        start_date = end_date - timedelta(days=6)
+        
+        # In real implementation:
+        # 1. Aggregate weekly metrics
+        # 2. Calculate KPIs and trends
+        # 3. Generate report document
+        # 4. Email to stakeholders
+        # 5. Update management dashboard
+        
+        report_data = {
+            "week_ending": end_date.isoformat(),
+            "total_trips": 0,
+            "total_distance_km": 0.0,
+            "average_utilization": 0.0,
+            "on_time_percentage": 0.0,
+            "fuel_efficiency": 0.0,
+            "incidents_count": 0
+        }
+        
+        logger.info(f"Weekly report generated for week ending {end_date}")
+        return {"status": "success", "report": report_data}
+        
+    except Exception as e:
+        logger.error(f"Weekly report generation failed: {e}")
+        raise self.retry(countdown=600)
+
+
+@celery_app.task(bind=True)
+def calculate_truck_performance_metrics(self):
+    """Calculate performance metrics for each truck"""
+    try:
+        logger.info("Calculating truck performance metrics")
+        
+        # In real implementation:
+        # 1. Calculate metrics per truck (utilization, efficiency, reliability)
+        # 2. Identify top and bottom performers
+        # 3. Update truck performance scores
+        # 4. Generate recommendations
+        
+        performance_data = {}  # Placeholder
+        
+        logger.info("Truck performance metrics calculated")
+        return {"status": "success", "truck_metrics": performance_data}
+        
+    except Exception as e:
+        logger.error(f"Truck performance calculation failed: {e}")
+        raise
+
+
+@celery_app.task(bind=True)
+def detect_route_anomalies(self):
+    """Detect unusual routes or deviations"""
+    try:
+        logger.info("Starting route anomaly detection")
+        
+        # In real implementation:
+        # 1. Analyze recent trip routes
+        # 2. Compare against historical patterns
+        # 3. Identify significant deviations
+        # 4. Flag potential issues (unauthorized stops, route inefficiencies)
+        # 5. Send alerts for critical anomalies
+        
+        anomalies = []  # Placeholder
+        
+        if anomalies:
+            logger.warning(f"Detected {len(anomalies)} route anomalies")
+            
+        logger.info("Route anomaly detection completed")
+        return {"status": "success", "anomalies_count": len(anomalies)}
+        
+    except Exception as e:
+        logger.error(f"Route anomaly detection failed: {e}")
+        raise