… .claude/CLAUDE.md
… .claude/hallucination-verification.md
… .claude/memory.md
… .cursor/toy-box/plan.md
… .cursor/toy-box/tracker.md
… .cursor/toy-box/validation-strategy.md
… README.md
… backend/services/ai_functions.py
… backend/services/ai_service.py
… backend/services/slack_service.py
… backend/tasks/notification_tasks.py
… backend/tasks/sync_tasks.py
… docs/GITHUB_ISSUES.md
… docs/O3_IMPLEMENTATION_PLAN.md
… scripts/env_test.py
… scripts/simple_loconav_test.py
… scripts/validate_loconav.py
… test_ai_final.py
… test_ai_service.py
… test_ai_simple.py
… test_final_working.py
… test_responses_api.py
… test_responses_direct.py
… test_responses_final.py
… test_responses_implementation.py
… test_slack_simple.py